refactor(mask): fix test error
